+++ microscope/stages/ludl.py
@@ -133,9 +133,6 @@ class _LudlController:
     def is_busy(self):
         pass
 
-    def been_homed(self):
-        return self.homed
-    
     def get_number_axes(self):
         return 2
     
@@ -348,15 +345,21 @@ class _LudlStage(microscope.abc.Stage):
         # Before a device can moved, it first needs to establish a
         # reference to the home position.  We won't be able to move
         # unless we home it first.
+        if not self.homed:
+            axes=self.axes
+            for axis in axes:
+                self.axes[axis].home()
+            self.homed = True
         return True
+
+
+    def may_move_on_enable(self) -> bool:
+        return not self.homed
+
 
     @property
     def axes(self) -> typing.Mapping[str, microscope.abc.StageAxis]:
         return self._axes
-
-    @property
-    def need_homed(self):
-        return not self.homed
 
     def move_by(self, delta: typing.Mapping[str, float]) -> None:
         """Move specified axes by the specified distance. """
@@ -375,13 +378,6 @@ class _LudlStage(microscope.abc.Stage):
             )
         self._dev_conn.wait_until_idle()
 
-    def home(self):
-        if self.need_homed:
-            axes=self.axes
-            for axis in axes:
-                self.axes[axis].home()
-            self.homed = True
-                
 #    def assert_filterwheel_number(self, number: int) -> None:
 #        assert number > 0 and number < 4