+++ NEWS.rst
@@ -13,6 +13,11 @@ Version 0.7.0 (upcoming)
     `Camera.get_trigger_type` does not return the same as
     `Camera.trigger_type` property.
 
+* Changes to device ABCs:
+
+  * The `Stage` ABC has a new method `may_move_on_enable` to hint
+    whether calling `enable` will cause the stage to move.
+
 * New devices supported:
 
   * Toptica iChrome MLE